test: optimize test files lookup
If you use nix flakes and enabling it using direnv, there will be
`.devenv` and `.direnv` directories in your root project directory. By
default vitest will try to look for test files in all directories unless
it is being excluded. Unfortunately `.devenv` and `.direnv` is not on
that list.

All we need is just tell vitest only to look for in `tests` directory.