refactor: don't use alias to export print function
The modular print function has signature
```js
const p = (node, path, print, options) => {
    // function body here
}
```

It is easy to recognise which is the main function responsibe to build
the doc. Unfortunately it has to be aliased when exported to prevent
name collision. And it is not obvious where the callstack come from when
debugging since it share the same name across multiple file/module.